Update zu solarswarm_setup - Skripte haben zusätzliche Checks, Kommentare und geben erste Log-Nachrichten an LOG_OUT (Konstante in den jeweiligen Skripten) aus - Bis auf die automatische Verwaltung der Manager wurde alles getestet. Jedoch wurden nicht alle Sonderfälle abgedeckt. Beispielsweise konnte für das Abfangen von Fehlern beim Hinzufügen von illegalen Labels nicht getestet werden, da keine illegalen Labels gefunden werden konnten (würde Spiel oder Recherche erfordern).
Fehlt noch:
- Weiterführen der Anleitung (inklusive Notizen aus Code-Kommentaren).
- Hochladen einiger Screenshots von Ausgaben.
- Datei mit Schema für "docker service create"-Statements und aufrufen diese in docker_init.bash.

+++ solarswarm_setup/system services/rx_copy.bash
@@ -1,43 +1,56 @@
 #!/bin/bash
+source /etc/environment
 COPY_TO_SSH=true
 AUTO_UPDATE_SSH=true
 # SWARM_MODE=true
-SW_SETUP=~/solarswarm_setup/
+SW_SETUP=/home/$MESH_IDENTITY/solarswarm_setup
+LOG_OUT=/dev/stdout # see docker_init.bash
 
 # update ~/.ssh/ as hosts can not directy copy into this directory
 if [ $COPY_TO_SSH == true ]; then
     if [ -d ~/.ssh/ ] && [ -d $SW_SETUP/rx/ssh/ ]; then
+        echo "[rx_copy] Checking '$SW_SETUP/rx/ssh/'" >>$LOG_OUT
         changed=false
         # move names, names_with_ip, config, and public keys to ssh_identities/
         if [ -f $SW_SETUP/rx/ssh/names ]; then
             mv $SW_SETUP/rx/ssh/names $SW_SETUP/ssh_identities/
+            echo "[rx_copy] Moved 'names' to '$SW_SETUP/ssh_identities/'" >>$LOG_OUT
             changed=true
         fi
 
         if [ -f $SW_SETUP/rx/ssh/names_with_ip ]; then
             mv $SW_SETUP/rx/ssh/names_with_ip $SW_SETUP/ssh_identities/
+            echo "[rx_copy] Moved 'names_with_ip' to '$SW_SETUP/ssh_identities/'" >>$LOG_OUT
             changed=true
         fi
 
         if [ -f $SW_SETUP/rx/ssh/config ]; then
             mv $SW_SETUP/rx/ssh/config $SW_SETUP/ssh_identities/
+            echo "[rx_copy] Moved 'config' to '$SW_SETUP/ssh_identities/'" >>$LOG_OUT
             changed=true
         fi
-
-        for pub_key in $(ls $SW_SETUP/rx/ssh/ | grep .pub); do
-            if [ ! $pub_key == $MESH_IDENTITY.pub ]; then
+        
+        for name in $($SW_SETUP/ssh_identities/names); do
+            if [ -f $SW_SETUP/rx/ssh/$name.pub ] && [ ! $pub_key == $MESH_IDENTITY.pub ]; then
                 mv $SW_SETUP/rx/ssh/$pub_key $SW_SETUP/ssh_identities/keys/
+                echo "[rx_copy] Moved '$pub_key' to '$SW_SETUP/ssh_identities/keys/'" >>$LOG_OUT
                 changed=true
             fi
         done
 
         # copy all to .ssh/ if anything has changed
         # done after everything was moved so ssh_identities/ and .ssh/ are synchronized
-        if [ $changed == true ] && [ $AUTO_UPDATE_SSH == true ]; then
-            # sudo cp $SW_SETUP/ssh_identities/names ~/.ssh/
-            # sudo cp $SW_SETUP/ssh_identities/names_with_ip ~/.ssh/
-            sudo cp $SW_SETUP/ssh_identities/config ~/.ssh/
-            sudo cp $SW_SETUP/ssh_identities/keys/*.pub ~/.ssh/ # all public keys
+        if [ ! -z $MESH_IP ] && [ $changed == true ] && [ $AUTO_UPDATE_SSH == true ]; then
+            cp $SW_SETUP/ssh_identities/config ~/.ssh/
+            sudo sed -i "s/own_name/$MESH_IP/" ~/.ssh/config # replace own_name with MESH_IP
+            for name in $($SW_SETUP/ssh_identities/names); do
+                if [ -f $SW_SETUP/ssh_identities/keys/$name.pub ]; then
+                    ssh-copy-id -i $SW_SETUP/ssh_identities/keys/$name.pub $MESH_IDENTITY@$MESH_IP
+                    # used to add public keys to '~/.ssh/' and '~/.ssh/authorized_keys/'
+                fi
+            done
+            # easier (but less safe): for pub_key in $(ls $SW_SETUP/ssh_identities/keys/); do...
+            echo "[rx_copy] Copied 'config' and public keys to '~/.ssh/'" >>$LOG_OUT
         fi
     fi
 fi
